requirements.txt
Django~=3.1.2
psycopg2-binary~=2.8.6
pandas~=1.1.4
djangorestframework~=3.12.1